requirements.txt
streamlit
pillow
pymupdf